Fixed a command line parsing bug that was preventing the fgetopts_long function from recognizing the long options correctly.

+++ src/fgetopt.c
@@ -82,7 +82,7 @@ int fgetopt_long (
 
                 for (op = longopts; op->name; ++op)
                 {
-                    if (((unsigned int) (end - begin)) == strlen (op->name)
+                    if (end - begin == strlen (op->name)
                         && strncmp (begin, op->name, end - begin) == 0)
                     {
                         int ret = 0;
